backend/requirements.txt
python-docx
docx2txt
PyMuPDF
pillow
pytesseract
flask
flask-cors
python-dotenv
python-telegram-bot
supabase
openai
youtube-transcript-api
edoc
httpx[socks]